Refactor ScarfPlot component imports and utilities for improved organization
This commit updates the import paths for various components and utilities within the ScarfPlot module, moving them to a more structured location under the utils directory. The changes include adjustments to the imports in ScarfPlot, ScarfPlotFigure, ScarfPlotSelectGroup, ScarfPlotSelectStimulus, and ScarfPlotSelectTimeline components. Additionally, new utility files for transformations and tooltip services have been added, enhancing code maintainability and clarity in the project's structure.
