+++ src/lib/plots/scarf/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './scarfPlotTransformations'
+export * from './transformations'
