+++ src/lib/plots/scarf/utils/tooltip.ts
