+++ src/lib/plots/transition-matrix/utils/transformations.ts
